tests/wpt/web-platform-tests/old-tests/submission/Microsoft/dragdrop/dragdrop_008.htm

<!doctype html>
<html>
    <head>
        <title>HTML5 Drag and Drop: True state of draggable attribute</title>
        <meta content="text/html; charset=UTF-8" http-equiv="Content-Type"/>
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/"/>
        <link rel="help" href="http://dev.w3.org/html5/spec/dnd.html#the-draggable-attribute"/>
        <meta name="assert" content="True state of draggable attribute"/>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
        <div id=log></div>
        <a href="#" style="display:none" id="target" draggable="true">Drag this</a>
        <script type="text/javascript">
            test(function() {assert_true(document.getElementById("target").draggable)}, "True state of draggable attribute");
        </script>
    </body>
</html>
